Adjust entity mapping; introduce sequence

+++ chucknorris-service/src/main/java/chuck/norris/repository/JokeRepository.java
@@ -11,6 +11,6 @@ import de.itasgmbh.workshop.model.Joke;
  * @author vkutsenko
  *
  */
-public interface JokeRepository extends JpaRepository<Joke, Integer> {
+public interface JokeRepository extends JpaRepository<Joke, Long> {
 
 }
